backend/requirements.txt
fastapi[all]
uvicorn
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pydantic
requests
google-auth
httpx
psycopg2-binary
sqlalchemy
boto3
pywebpush
supabase
alembic
aiofiles
pyotp
qrcode
Pillow
celery
redis
google-cloud-storage
